Refactor modals folder location
