+++ src/components/modals/home/StepOneLogin.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
-import Button from "../commons/Button";
-import { IModal, IOpenModal } from "../../types/types";
+import Button from "../../commons/Button";
+import { IModal, IOpenModal } from "../../../types/types";
 import { Checkbox } from "antd";
-import closeModalHandler from "../../actions/closeModalHandler";
-import openModalHandler from "../../actions/openModalHandler";
+import closeModalHandler from "../../../actions/closeModalHandler";
+import openModalHandler from "../../../actions/openModalHandler";
 
 const StepOneLogin: React.FC<IModal> = ({
   content,
