+++ src/components/modals/wallet/SendTransaction.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 import { Modal } from "antd";
-import Button from "../commons/Button";
-import { IModalSendTransaction } from "../../types/types";
+import Button from "../../commons/Button";
+import { IModalSendTransaction } from "../../../types/types";
 import { Input, Tooltip } from "antd";
 import { InfoCircleOutlined, UserOutlined } from "@ant-design/icons";
-import closeModalHandler from "../../actions/closeModalHandler";
+import closeModalHandler from "../../../actions/closeModalHandler";
 
 const SendTransaction: React.FC<IModalSendTransaction> = ({
   content,
